Added octree implementation of optical depth finder. Cleaned up.

+++ include/Vec.h
@@ -19,25 +19,25 @@
 #include "Definitions.h"
 
 struct Vec3;
-Vec3 operator*(double r, const Vec3 &v);
+Vec3 operator*(FLOAT r, const Vec3 &v);
 
 struct Vec3 {
   union {
     struct {
-      double x;
-      double y;
-      double z;
+      FLOAT x;
+      FLOAT y;
+      FLOAT z;
     };
-    double D[3];
+    FLOAT D[3];
   };
 
   Vec3() {}
 
-  Vec3(double _x, double _y, double _z) : x(_x), y(_y), z(_z) {}
+  Vec3(FLOAT _x, FLOAT _y, FLOAT _z) : x(_x), y(_y), z(_z) {}
 
-  double &operator[](unsigned int i) { return D[i]; }
+  FLOAT &operator[](unsigned int i) { return D[i]; }
 
-  const double &operator[](unsigned int i) const { return D[i]; }
+  const FLOAT &operator[](unsigned int i) const { return D[i]; }
 
   Vec3 operator+(const Vec3 &r) const {
     return Vec3(x + r.x, y + r.y, z + r.z);
@@ -47,13 +47,13 @@ struct Vec3 {
     return Vec3(x - r.x, y - r.y, z - r.z);
   }
 
-  Vec3 operator*(double r) const { return Vec3(x * r, y * r, z * r); }
+  Vec3 operator*(FLOAT r) const { return Vec3(x * r, y * r, z * r); }
 
   Vec3 operator^(const Vec3 &r) const {
     return Vec3(y * r.z - z * r.y, z * r.x - x * r.z, x * r.y - y * r.x);
   }
 
-  Vec3 operator/(double r) const { return Vec3(x / r, y / r, z / r); }
+  Vec3 operator/(FLOAT r) const { return Vec3(x / r, y / r, z / r); }
 
   Vec3 cmul(const Vec3 &r) const { return Vec3(x * r.x, y * r.y, z * r.z); }
 
@@ -73,25 +73,25 @@ struct Vec3 {
     return *this;
   }
 
-  Vec3 &operator*=(double r) {
+  Vec3 &operator*=(FLOAT r) {
     x *= r;
     y *= r;
     z *= r;
     return *this;
   }
 
-  Vec3 &operator/=(double r) {
+  Vec3 &operator/=(FLOAT r) {
     x /= r;
     y /= r;
     z /= r;
     return *this;
   }
 
-  double Norm() const { return sqrt(x * x + y * y + z * z); }
+  FLOAT Norm() const { return sqrt(x * x + y * y + z * z); }
 
-  double Norm2() const { return sqrt(x * x + y * y); }
+  FLOAT Norm2() const { return sqrt(x * x + y * y); }
 
-  double NormSquared() const { return x * x + y * y + z * z; }
+  FLOAT NormSquared() const { return x * x + y * y + z * z; }
 
   Vec3 Normalized() const { return *this / Norm(); }
 
@@ -106,6 +106,6 @@ struct Vec3 {
   }
 };
 
-inline Vec3 operator*(double r, const Vec3 &v) {
+inline Vec3 operator*(FLOAT r, const Vec3 &v) {
   return Vec3(v.x * r, v.y * r, v.z * r);
 }
